requirements.txt
celery==5.6.2
Django==5.2.11
redis==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
